backend/requirements.txt
# Django Framework
Django==4.2.27

# Environment variables
python-dotenv==1.0.0

# CORS headers for cross-origin requests
django-cors-headers==4.3.1

# Excel file handling
openpyxl==3.1.2

# Google Gemini AI SDK
google-genai==0.2.2

